+++ server/src/macroLanguageService/services/macroCommands.ts
@@ -40,34 +40,32 @@ export class MacroCommand {
 					return true;
 				});
 
-
 				func.accept(candidate => {
-					if (candidate.type === nodes.NodeType.SequenceNumber) {
-						const nnode = (<nodes.SequenceNumber>candidate).getNumber();
-						const number = nnode?.getText().toLocaleLowerCase().split('n').pop();
-						const labels = gotoLabelList.filter(a => a.getText() === number);
-
-						if (nnode) {
-							const n = 'N' + seq;
-							const start = document.positionAt(nnode.offset);
-							const end = document.positionAt(nnode.end);
-							textEdits.push(TextEdit.del(Range.create(start, end)));
-							textEdits.push(TextEdit.insert(start, n));
-
-							for (const label of labels){
-								const start = document.positionAt(label.offset);
-								const end = document.positionAt(label.end);
+					this.skip(candidate, () => {
+						if (candidate.type === nodes.NodeType.SequenceNumber) {
+							const nnode = (<nodes.SequenceNumber>candidate).getNumber();
+							const labels = gotoLabelList.filter(a => a.getText() === nnode?.getText());
+	
+							if (nnode) {
+								const start = document.positionAt(nnode.offset);
+								const end = document.positionAt(nnode.end);
 								textEdits.push(TextEdit.del(Range.create(start, end)));
-								textEdits.push(TextEdit.insert(start, seq.toString()));
-								var index = gotoLabelList.indexOf(label);
-								if (index > -1) {
-									gotoLabelList.splice(index, 1);
+								textEdits.push(TextEdit.insert(start, seq+''));
+	
+								for (const label of labels) {
+									const start = document.positionAt(label.offset);
+									const end = document.positionAt(label.end);
+									textEdits.push(TextEdit.del(Range.create(start, end)));
+									textEdits.push(TextEdit.insert(start, seq.toString()));
+									var index = gotoLabelList.indexOf(label);
+									if (index > -1) {
+										gotoLabelList.splice(index, 1);
+									}
 								}
+								seq = seq + inc;
 							}
-
-							seq = seq + inc;
 						}
-					}
+					});
 					return true;
 				});
 				if (textEdits.length > 0) {
@@ -100,16 +98,18 @@ export class MacroCommand {
 			if (func) {
 				const textEdits:TextEdit[] = [];
 				func.accept(candidate => {
-					if (candidate.type === nodes.NodeType.Statement && !candidate.findAParent(nodes.NodeType.SequenceNumber)) {
-						const statement = (<nodes.NcStatement>candidate);
-						if (statement) {
-							const n = 'N' + seq + ' ';
-							const start = document.positionAt(statement.offset);
-							textEdits.push(TextEdit.insert(start, n));
-							seq = seq + inc;
-							return false;
+					this.skip(candidate, () => {
+						if (candidate.type === nodes.NodeType.Statement && !candidate.findAParent(nodes.NodeType.SequenceNumber)) {
+							const statement = (<nodes.NcStatement>candidate);
+							if (statement) {
+								const n = 'N' + seq + ' ';
+								const start = document.positionAt(statement.offset);
+								textEdits.push(TextEdit.insert(start, n));
+								seq = seq + inc;
+								return false;
+							}
 						}
-					}
+					});
 					return true;
 				});
 
@@ -129,16 +129,46 @@ export class MacroCommand {
 	private getMaxSequenceNumber(node: nodes.Node) : number {
 		let seq = -1;
 		node.accept(candidate => {
-			if (candidate.type === nodes.NodeType.SequenceNumber) {
-				const nnode = (<nodes.SequenceNumber>candidate).getNumber();
-				const number = Number(nnode?.getText().toLocaleLowerCase().split('n').pop());
-				if (nnode && !Number.isNaN(number)) {
-					seq = Math.max(seq, number);
-					return false;
+			this.skip(candidate, () => {
+				if (candidate.type === nodes.NodeType.SequenceNumber) {
+					const nnode = (<nodes.SequenceNumber>candidate).getNumber();
+					const number = Number(nnode?.getText().toLocaleLowerCase().split('n').pop());
+					if (nnode && !Number.isNaN(number)) {
+						seq = Math.max(seq, number);
+						return false;
+					}
 				}
-			}
+			});
 			return true;
 		});
 		return seq;
+	}
+
+	private _skip = false;
+	private skip(candidate:nodes.Node, f:() => void) {
+		function checkSkip(candidate:nodes.Node, code:string) : boolean {
+			let result = false;
+			for (const child of candidate.getChildren()) {
+				result =  checkSkip(child, code);
+				const childText = child.getText().toLocaleLowerCase();
+				if (childText === code) {
+					result = true;
+					return result;
+				}
+			} 
+			return result;
+		}
+
+		if (candidate.type === nodes.NodeType.Statement || candidate.type === nodes.NodeType.SequenceNumber) {
+			if (checkSkip(candidate, 'g11')) {
+				this._skip = false;
+			}
+			if (!this._skip) {
+				f();
+			}
+			if (checkSkip(candidate, 'g10')) {
+				this._skip = true;
+			}
+		}
 	}
 }